Added an example of how to intercept exceptions due to incorrect user queries
When the users enters an incorrect query, _Tire_ raises a `Tire::Search::SearchRequestFailed` exception.

It is unfortunate to propagate the raw exception to the page, and display the “Something went wrong” page
in this case -- it's better to display the default result set, or the search form, and indicate to the
user her query was incorrect.

This commit contains a trivial implementation via the _Rails'_ `rescue_from` method.

See <http://github.com/karmi/tire/issues/164#issuecomment-3096189>.